Change Salary type to double

+++ attendance_system_backend/Models/Employee.cs
@@ -39,7 +39,7 @@ namespace attendance_system_backend.Models
         public string Position { get; set; }
 
         [Range(0, double.MaxValue, ErrorMessage = "Salary must be a non-negative value.")]
-        public decimal Salary { get; set; }
+        public double Salary { get; set; }
 
         [Required(ErrorMessage = "Gender is required.")]
         public Gender Gender { get; set; }
